spark/notebooks/iot_data_analysis/requirements.txt
pyspark==3.5.3
kafka-python==2.0.2
pandas
psycopg2-binary
